built-plex/requirements.txt
requests
tzlocal
plexapi
PyYaml
requests
